Summary: Added new entities and repositories Issue link: https://gitlab.com/kv.061.java/project_kv_061/issues/3 Changes: Added entities : Actor, ActorRole, Order, OrderDetails, OrderEvent, OrderEventType and repositories for them Changes datatype of id to Long

+++ src/main/java/com/softserve/itacademy/kek/repositories/GlobalPropertiesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GlobalPropertiesRepository extends CrudRepository<GlobalProperties, Integer> {
+public interface GlobalPropertiesRepository extends CrudRepository<GlobalProperties, Long> {
 }
